Refactor : test에 있는 controller 관련 테스트들 Controller 패키지에 이동
