+++ src/test/java/lotto/Controller/PlayTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Controller;
 
 import lotto.model.Lotto;
 import org.assertj.core.api.Assertions;
